Refactor to more classes
