+++ src/Stryker.Core/Stryker.Core/Reporters/Progress/IProgressBarReporter.cs
@@ -1,0 +1,6 @@
+﻿using System;
+
+namespace Stryker.Core.Reporters.Progress
+{
+    
+}
